test: enforce single cfg test module per file
