+++ crates/monochange/src/__tests__/publish_readiness_tests.rs
@@ -1,6 +1,84 @@
 #![allow(clippy::disallowed_methods)]
 use super::*;
 
+async fn validate_publish_readiness_artifact(
+	root: &Path,
+	configuration: &WorkspaceConfiguration,
+	prepared_release: Option<&PreparedRelease>,
+	selected_packages: &BTreeSet<String>,
+	artifact_path: &Path,
+) -> MonochangeResult<()> {
+	let artifact = read_report_artifact(artifact_path)?;
+	let current_report = build_publish_readiness_report_for_publish(
+		root,
+		configuration,
+		prepared_release,
+		selected_packages,
+	)
+	.await?;
+	validate_publish_readiness_report(&artifact, &current_report)
+}
+
+fn validate_publish_readiness_report(
+	artifact: &PublishReadinessReport,
+	current: &PublishReadinessReport,
+) -> MonochangeResult<()> {
+	validate_readiness_artifact_header(artifact)?;
+	validate_readiness_artifact_status(artifact)?;
+	validate_readiness_current_status(current)?;
+	validate_readiness_release_commit(artifact, current)?;
+	validate_readiness_input_fingerprint(artifact, current)?;
+	validate_readiness_packages(artifact, current)
+}
+
+fn validate_readiness_artifact_status(report: &PublishReadinessReport) -> MonochangeResult<()> {
+	if report.status == PublishReadinessGlobalStatus::Ready {
+		return Ok(());
+	}
+	Err(MonochangeError::Config(
+		"publish readiness artifact is blocked; rerun `monochange step publish-readiness` and resolve blockers before `monochange step publish-packages`".to_string(),
+	))
+}
+
+fn validate_readiness_current_status(report: &PublishReadinessReport) -> MonochangeResult<()> {
+	if report.status == PublishReadinessGlobalStatus::Ready {
+		return Ok(());
+	}
+	Err(MonochangeError::Config(
+		"current publish readiness is blocked; rerun `monochange step publish-readiness` and resolve blockers before `monochange step publish-packages`".to_string(),
+	))
+}
+
+fn validate_readiness_packages(
+	artifact: &PublishReadinessReport,
+	current: &PublishReadinessReport,
+) -> MonochangeResult<()> {
+	let artifact_packages = package_identities(&artifact.packages)?;
+	let current_packages = package_identities(&current.packages)?;
+	if artifact.package_set_fingerprint != package_set_fingerprint(&artifact.packages) {
+		return Err(MonochangeError::Config(
+			"publish readiness artifact package fingerprint does not match its package list"
+				.to_string(),
+		));
+	}
+	if artifact_packages == current_packages {
+		return Ok(());
+	}
+	let missing = current_packages
+		.difference(&artifact_packages)
+		.map(render_package_identity)
+		.collect::<Vec<_>>();
+	let stale = artifact_packages
+		.difference(&current_packages)
+		.map(render_package_identity)
+		.collect::<Vec<_>>();
+	Err(MonochangeError::Config(format!(
+		"publish readiness artifact package set is stale or does not match selected packages (missing: {}; stale: {})",
+		render_package_identity_list(&missing),
+		render_package_identity_list(&stale)
+	)))
+}
+
 fn sample_publish_outcome(
 	status: package_publish::PackagePublishStatus,
 ) -> package_publish::PackagePublishOutcome {
