+++ crates/monochange_core/src/__tests__/mod_tests.rs
@@ -1,0 +1,5 @@
+#[path = "lib_tests.rs"]
+mod lib_tests;
+
+#[path = "sync_tests.rs"]
+mod sync_tests;
