+++ crates/monochange_core/src/lib.rs
@@ -5577,9 +5577,5 @@ pub mod schema {
 }
 
 #[cfg(test)]
-#[path = "__tests__/sync_tests.rs"]
-mod sync_tests;
-
-#[cfg(test)]
-#[path = "__tests__/lib_tests.rs"]
+#[path = "__tests__/mod_tests.rs"]
 mod tests;
