+++ crates/monochange_dart/src/__tests__/mod_tests.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[path = "benchmark_tests.rs"]
+mod benchmark_tests;
+
+#[path = "lib_tests.rs"]
+mod lib_tests;
+
+#[path = "sync_tests.rs"]
+mod sync_tests;
