+++ crates/monochange_npm/src/__tests__/mod_tests.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+#[path = "lib_tests.rs"]
+mod lib_tests;
+
+#[path = "sync_tests.rs"]
+mod sync_tests;
