+++ crates/monochange_npm/src/lib.rs
@@ -1302,9 +1302,5 @@ fn released_versions_by_package_id(
 }
 
 #[cfg(test)]
-#[path = "__tests__/sync_tests.rs"]
-mod sync_tests;
-
-#[cfg(test)]
-#[path = "__tests__/lib_tests.rs"]
+#[path = "__tests__/mod_tests.rs"]
 mod tests;
